[FE] REFACT: 프로필 편집 컴포넌트화 #70
